ide_completion

Module completions

This module defines an accumulator for completions which are going to be presented to user.

Modules

attribute ğŸ”’

Completion for (built-in) attributes, derives and lints.

dot ğŸ”’

Completes references after dot (fields and method calls).

env_vars ğŸ”’

Completes environment variables defined by Cargo (https://doc.rust-lang.org/cargo/reference/environment-variables.html)

expr ğŸ”’

Completion of names from the current scope in expression position.

extern_abi ğŸ”’

Completes function abi strings.

extern_crate ğŸ”’

Completion for extern crates

field ğŸ”’

Completion of field list position.

flyimport ğŸ”’

See import_on_the_fly.

fn_param ğŸ”’

See complete_fn_param.

format_string ğŸ”’

Completes identifiers in format string literals.

item_list ğŸ”’

Completion of paths and keywords at item list position.

keyword ğŸ”’

Completes where and for keywords.

lifetime ğŸ”’

Completes lifetimes and labels.

macro_def ğŸ”’

Completion for macro meta-variable segments

mod_ ğŸ”’

Completes mod declarations.

pattern ğŸ”’

Completes constants and paths in unqualified patterns.

postfix ğŸ”’

Postfix completions, like Ok(10).ifl$0 => if let Ok() = Ok(10) { $0 }.

ra_fixture ğŸ”’

Injected completions for #[rust_analyzer::rust_fixture].

record ğŸ”’

Complete fields in record literals and patterns.

snippet ğŸ”’

This file provides snippet completions, like pd => eprintln!(...).

type ğŸ”’

Completion of names from the current scope in type position.

use_ ğŸ”’

Completion for use trees

vis ğŸ”’

Completion for visibility specifiers.

Structs

Completions

Represents an in-progress set of completions being built.

Functions

complete_name ğŸ”’

complete_name_ref ğŸ”’

complete_patterns ğŸ”’

enum_variants_with_paths ğŸ”’

Calls the callback for each variant of the provided enum with the path to the variant. Skips variants that are visible with single segment paths.