ide_completion::completions::item_list

Module trait_impl

Completion for associated items in a trait implementation.

This module adds the completion items related to implementing associated items within an impl Trait for Struct block. The current context node must be within either a FN, TYPE_ALIAS, or CONST node and an direct child of an IMPL.

Examples

Considering the following trait impl:

trait SomeTrait {
    fn foo();
}

impl SomeTrait for () {
    fn f$0
}

may result in the completion of the following method:

impl SomeTrait for () {
    fn foo() {}$0
}

Enums

AsyncSugaring ğŸ”’

ImplCompletionKind ğŸ”’

Functions

add_const_impl ğŸ”’

add_function_impl ğŸ”’

add_function_impl_ ğŸ”’

add_type_alias_impl ğŸ”’

complete_trait_impl ğŸ”’

complete_trait_impl_const ğŸ”’

complete_trait_impl_fn ğŸ”’

complete_trait_impl_item_by_name ğŸ”’

complete_trait_impl_name ğŸ”’

complete_trait_impl_type_alias ğŸ”’

function_declaration ğŸ”’

get_transformed_assoc_item ğŸ”’

Transform a relevant associated item to inline generics from the impl, remove attrs and docs, etc.

get_transformed_fn ğŸ”’

Transform a relevant associated item to inline generics from the impl, remove attrs and docs, etc.

make_const_compl_syntax ğŸ”’