ide_completion

Module context

See CompletionContext structure.

Modules

analysis 🔒

Module responsible for analyzing the code surrounding the cursor for completion.

Structs

AttrCtx 🔒

CompletionContext 🔒

CompletionContext is created early during completion to figure out, where exactly is the cursor, syntax-wise.

DotAccess 🔒

Information about the field or method access we are completing.

DotAccessExprCtx 🔒

LifetimeContext 🔒

The state of the lifetime we are completing.

NameContext 🔒

The state of the name we are completing.

NameRefContext 🔒

The state of the NameRef we are completing.

ParamContext 🔒

PathCompletionCtx 🔒

The state of the path we are currently completing.

PathExprCtx 🔒

PatternContext 🔒

The state of the pattern we are completing.

QualifierCtx 🔒

Existing qualifiers for the thing we are currently completing.

Enums

BreakableKind 🔒

CompleteSemicolon 🔒

CompletionAnalysis 🔒

The identifier we are currently completing.

DotAccessKind 🔒

ItemListKind 🔒

The kind of item list a PathKind::Item belongs to.

LifetimeKind 🔒

The kind of lifetime we are completing.

NameKind 🔒

The kind of the name we are completing.

NameRefKind 🔒

The kind of the NameRef we are completing.

ParamKind 🔒

PathKind 🔒

The kind of path we are completing right now.

PatternRefutability 🔒

Qualified 🔒

TypeAscriptionTarget 🔒

TypeLocation 🔒

Original file ast nodes

Visible 🔒

Constants

COMPLETION_MARKER 🔒

INTERNAL_FEATURES_LIST 🔒

OP_TRAIT_LANG 🔒

Statics

INTERNAL_FEATURES 🔒

Type Aliases

ExistingDerives 🔒