ide_completion::context

Module analysis

Module responsible for analyzing the code surrounding the cursor for completion.

Structs

AnalysisResult 🔒

ExpansionResult 🔒

Functions

analyze 🔒

Fill the completion context, this is what does semantic reasoning about the surrounding context of the completion location.

ancestors_in_file_compensated 🔒

classify_lifetime 🔒

classify_name 🔒

classify_name_ref 🔒

expand 🔒

expand_and_analyze 🔒

expand_maybe_stop 🔒

Expand attributes and macro calls at the current cursor position for both the original file and fake file repeatedly. As soon as one of the two expansions fail we stop so the original and speculative states stay in sync.

expected_type_and_name 🔒

Calculate the expected type and name of the cursor position.

fetch_immediate_impl_or_trait 🔒

find_node_in_file 🔒

Attempts to find node inside syntax via node’s text range. If the fake identifier has been inserted after this node or inside of this node use the _compensated version instead.

find_node_in_file_compensated 🔒

Attempts to find node inside syntax via node’s text range while compensating for the offset introduced by the fake ident. This is wrong if node comes before the insertion point! Use find_node_in_file instead.

find_opt_node_in_file 🔒

Attempts to find node inside syntax via node’s text range. If the fake identifier has been inserted after this node or inside of this node use the _compensated version instead.

find_opt_node_in_file_compensated 🔒

Attempts to find node inside syntax via node’s text range while compensating for the offset introduced by the fake ident.. This is wrong if node comes before the insertion point! Use find_node_in_file instead.

has_in_newline_expr_first 🔒

Similar to has_parens, heuristic sensing incomplete statement before ambiguous Expr

has_parens 🔒

When writing in the middle of some code the following situation commonly occurs (| denotes the cursor):

is_after_if_expr 🔒

is_in_block 🔒

is_in_breakable 🔒

is_in_token_of_for_loop 🔒

left_ancestors 🔒

next_non_trivia_sibling 🔒

next_non_trivia_token 🔒

path_or_use_tree_qualifier 🔒

pattern_context_for 🔒

prev_special_biased_token_at_trivia 🔒

token_at_offset_ignore_whitespace 🔒