ide_completion::context::analysis

Function has_parens

fn has_parens(node: &dyn HasArgList) -> bool

When writing in the middle of some code the following situation commonly occurs (| denotes the cursor):

value.method|
(1, 2, 3)

Here, we want to complete the method parentheses & arguments (if the corresponding settings are on), but the thing is parsed as a method call with parentheses. Therefore we use heuristics: if the parentheses are on the next line, consider them non-existent.