ide_completion

Module render

render module provides utilities for rendering completion suggestions into code pieces that will be presented to user.

Modules

const_ 🔒

Renderer for const fields.

function 🔒

Renderer for function calls.

literal 🔒

Renderer for enum variants.

macro_ 🔒

Renderer for macro invocations.

pattern 🔒

Renderer for patterns.

type_alias 🔒

Renderer for type aliases.

union_literal 🔒

Renderer for union literals.

variant 🔒

Code common to structs, unions, and enum variants.

Structs

RenderContext 🔒

Interface for data and methods required for items rendering.

Functions

adds_ret_type_arrow 🔒

compute_exact_name_match 🔒

compute_has_local_inherent_impl 🔒

compute_ref_match 🔒

compute_type_match 🔒

field_with_receiver 🔒

get_import_name 🔒

match_types 🔒

path_ref_match 🔒

render_expr 🔒

render_field 🔒

render_path_resolution 🔒

render_pattern_resolution 🔒

render_resolution_pat 🔒

render_resolution_path 🔒

render_resolution_simple_ 🔒

render_resolution_with_import 🔒

render_resolution_with_import_pat 🔒

render_tuple_field 🔒

render_type_inference 🔒

render_type_keyword_snippet 🔒

res_to_kind 🔒

scope_def_docs 🔒

scope_def_is_deprecated 🔒

scope_def_to_name 🔒