Crate ide_completion

completions crate provides utilities for generating completions of user input.

Structs

CompletionConfig

CompletionFieldsToResolve

CompletionItem

CompletionItem describes a single completion entity which expands to 1 or more entries in the editor pop-up.

CompletionRelevance

Snippet

A user supplied snippet.

Enums

AutoImportExclusionType

CallableSnippets

CompletionItemKind

The type of the completion item.

CompletionItemRefMode

CompletionRelevancePostfixMatch

CompletionRelevanceReturnType

CompletionRelevanceTypeMatch

SnippetScope

A snippet scope describing where a snippet may apply to. These may differ slightly in meaning depending on the snippet trigger.

Functions

completions

Main entry point for completion. We run completion as a two-phase process.

resolve_completion_edits

Resolves additional completion data at the position given. This is used for import insertion done via completions like flyimport and custom user snippets.