Crate ide_completion

completions crate provides utilities for generating completions of user input.

Structs

CompletionConfig

CompletionItem

CompletionItem describes a single completion variant in the editor pop-up. It is basically a POD with various properties. To construct a CompletionItem, use new method and the Builder struct.

CompletionRelevance

Snippet

A user supplied snippet.

Enums

CallableSnippets

CompletionItemKind

The type of the completion item.

CompletionRelevancePostfixMatch

SnippetScope

A snippet scope describing where a snippet may apply to. These may differ slightly in meaning depending on the snippet trigger.

Functions

completions

Main entry point for completion. We run completion as a two-phase process.

resolve_completion_edits

Resolves additional completion data at the position given. This is used for import insertion done via completions like flyimport and custom user snippets.