Crate ide_completion

completions crate provides utilities for generating completions of user input.

Structs

CompletionConfig

CompletionItem
CompletionItem describes a single completion variant in the editor pop-up. It is basically a POD with various properties. To construct a CompletionItem, use new method and the Builder struct.

CompletionRelevance

ImportEdit
An extra import to add after the completion is applied.

Enums

CompletionItemKind

InsertTextFormat

Functions

completions
Main entry point for completion. We run completion as a two-phase process.

resolve_completion_edits
Resolves additional completion data at the position given.