ide_completion

Module context

See CompletionContext structure.

Modules

analysis ğŸ”’

Module responsible for analyzing the code surrounding the cursor for completion.

Structs

AttrCtx ğŸ”’

CompletionContext ğŸ”’

CompletionContext is created early during completion to figure out, where exactly is the cursor, syntax-wise.

DotAccess ğŸ”’

Information about the field or method access we are completing.

DotAccessExprCtx ğŸ”’

LifetimeContext ğŸ”’

The state of the lifetime we are completing.

NameContext ğŸ”’

The state of the name we are completing.

NameRefContext ğŸ”’

The state of the NameRef we are completing.

ParamContext ğŸ”’

PathCompletionCtx ğŸ”’

The state of the path we are currently completing.

PathExprCtx ğŸ”’

PatternContext ğŸ”’

The state of the pattern we are completing.

QualifierCtx ğŸ”’

Existing qualifiers for the thing we are currently completing.

Enums

BreakableKind ğŸ”’

CompleteSemicolon ğŸ”’

CompletionAnalysis ğŸ”’

The identifier we are currently completing.

DotAccessKind ğŸ”’

ItemListKind ğŸ”’

The kind of item list a PathKind::Item belongs to.

LifetimeKind ğŸ”’

The kind of lifetime we are completing.

NameKind ğŸ”’

The kind of the name we are completing.

NameRefKind ğŸ”’

The kind of the NameRef we are completing.

ParamKind ğŸ”’

PathKind ğŸ”’

The kind of path we are completing right now.

PatternRefutability ğŸ”’

Qualified ğŸ”’

TypeAscriptionTarget ğŸ”’

TypeLocation ğŸ”’

Original file ast nodes

Visible ğŸ”’

Constants

COMPLETION_MARKER ğŸ”’

OP_TRAIT_LANG ğŸ”’

Type Aliases

ExistingDerives ğŸ”’