ide_completion

Module render

render module provides utilities for rendering completion suggestions into code pieces that will be presented to user.

Modules

const_ ğŸ”’

Renderer for const fields.

function ğŸ”’

Renderer for function calls.

literal ğŸ”’

Renderer for enum variants.

macro_ ğŸ”’

Renderer for macro invocations.

pattern ğŸ”’

Renderer for patterns.

type_alias ğŸ”’

Renderer for type aliases.

union_literal ğŸ”’

Renderer for union literals.

variant ğŸ”’

Code common to structs, unions, and enum variants.

Structs

RenderContext ğŸ”’

Interface for data and methods required for items rendering.

Functions

compute_exact_name_match ğŸ”’

compute_ref_match ğŸ”’

compute_type_match ğŸ”’

field_with_receiver ğŸ”’

get_import_name ğŸ”’

match_types ğŸ”’

path_ref_match ğŸ”’

render_expr ğŸ”’

render_field ğŸ”’

render_path_resolution ğŸ”’

render_pattern_resolution ğŸ”’

render_resolution_pat ğŸ”’

render_resolution_path ğŸ”’

render_resolution_simple_ ğŸ”’

render_resolution_with_import ğŸ”’

render_resolution_with_import_pat ğŸ”’

render_tuple_field ğŸ”’

render_type_inference ğŸ”’

res_to_kind ğŸ”’

scope_def_docs ğŸ”’

scope_def_is_deprecated ğŸ”’

scope_def_to_name ğŸ”’